+++ OpenGL/src/scene/ClearColourScene.h
@@ -8,7 +8,7 @@ namespace scene {
 		ClearColourScene();
 		~ClearColourScene() override;
 
-		void OnUpdate(float deltaTime, GLFWwindow* window) override;
+		void OnUpdate(double deltaTime, GLFWwindow* window) override;
 		void OnRender() override;
 		void OnImGuiRender() override;
 	private:
